// types/forms/enums.ts
export enum TextTypeEnums {
  withBtnInside = 'withBtn',
}

export enum BusinessFormRubro {
  tecnologia = 'Tecnología',
  petroleo = 'Petróleo',
  consultoria = 'Consultoría',
}

export enum BusinessFormRango {
  n1n10 = '1-10',
  n11n50 = '11-50',
  n51n200 = '51-200',
  n201n500 = '201-500',
  mas500 = 'Más de 500',
}